requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.0.10
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.4
psycopg2-binary==2.9.3
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.0.0
